Version 1.1.1 Created Rcpp module IntegratorPOUMM, which replaces the previous function likPOUMM...C.
pruneTree function now receives z as a second parameter. The returned pruneInfo object contains the tree and z plus an Integrator object. The calculation of a, b, c is done by calling a method with this object.

Functions summary and plot are unlikely to work with previous versions.

Fixed bugs in specPOUMM.R
Added function fitted for a POUMM object.