Many fixes: Updated CITATION and References in user guide to include more recent versions of the our preprints/publications.
Fixed a problem with tarball size exceeding 5Mb: the stored fitMCMC member in the POUMM fit object was too big due to storing a field sampling.parameters in each mcmc chain object. After that, the size of the file UserGuideCache.RData in vignettes was reduced from 4.2Mb to 1.1Mb.

Changed default setting for gammaMCMC from 2/3 to 0.50001 - needed for faster convergence and for consistency with previous POUMM fits, which were run with gamma=0.5. This follows a change in the adaptMCMC package.

UPdated web-site.